fix: avoid passing a number that isnt an int in query params

+++ src/schemas/fizz-buzz.schemas.ts
@@ -2,8 +2,8 @@ import * as z from "zod";
 
 export const fizzBuzzQuerySchema = z
   .object({
-    from: z.coerce.number().min(1).max(100).default(1),
-    to: z.coerce.number().min(1).max(100).default(100),
+    from: z.coerce.number().int().min(1).max(100).default(1),
+    to: z.coerce.number().int().min(1).max(100).default(100),
   })
   .refine((data) => data.to >= data.from, {
     message: "'to' must be greater than or equal to 'from'",
